rename analysis.py to cluster.py
