+++ BALSAMIC/commands/options.py
@@ -1,7 +1,7 @@
 """Balsamic commands common options."""
 import click
 
-from BALSAMIC.constants.analysis import RunMode, ClusterProfile, QOS, ClusterMailType
+from BALSAMIC.constants.cluster import RunMode, ClusterProfile, QOS, ClusterMailType
 from BALSAMIC.constants.cache import GenomeVersion
 
 OPTION_GENOME_VERSION = click.option(
